add more test cases
